requirements.txt
anthropic>=0.19.1  # For calling Claude API
openai>=1.59.8     # Updated to new version
elasticsearch>=8.0.0
flask>=2.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
tqdm>=4.62.0